pyproject.toml
[tool.poetry]
name = "api-payments"
version = "0.1.0"
description = ""
authors = ["Vagner Bessa"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
tortoise-orm = "^0.24.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
pydantic-settings = "^2.7.1"
slowapi = "^0.1.9"
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"
aerich = "^0.8.1"
tomlkit = "^0.13.2"

[tool.poetry.dependencies.pydantic]
extras = ["email"]
version = "^2.10.6"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
asgi-lifespan = "^2.1.0"
async-asgi-testclient = "^1.4.11"
watchfiles = "^1.0.4"

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "app/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
